[FEAT] Players Info API #52
- Add API providing players information such as username, avatar
