+++ src/routes/playerRouter.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import playerController from '../controllers/playerController';
+
+const router = express.Router();
+
+router.get('/:roomId', playerController.getPlayersInfo);
+
+export default router;
