+++ swagger/swagger.yaml
@@ -23,6 +23,8 @@ tags:
     description: 주식 정보 등
   - name: room dashboard
     description: 게임 현황
+  - name: player
+    description: 게임하고 있는 유저들의 정보
 
 paths:
   /me:
@@ -344,6 +346,27 @@ paths:
                 type: array
                 items:
                   $ref: '#/components/schemas/Dashboard'
+  /player/{roomId}:
+    get:
+      tags:
+        - player
+      summary: players information
+      operationId: getPlayersInfo
+      parameters:
+        - name: roomId
+          in: path
+          required: true
+          schema:
+            type: integer
+      responses:
+        200:
+          description: successful operation
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/Player'
 
 components:
   schemas:
@@ -488,3 +511,12 @@ components:
                 format: float
               rank:
                 type: integer
+    Player:
+      type: object
+      properties:
+        userId:
+          type: number
+        username:
+          type: string
+        avatar:
+          type: string
